+++ seeker/src/main/java/se452/group9/seeker/repo/CompanyRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompanyRepository extends JpaRepository<Company, Integer> {
-    Company findBycompanyID(int id);
+public interface CompanyRepository extends JpaRepository<Company, Long> {
+    Company findBycompanyID(long id);
     List<Company> findBycompanyName(String companyName);
 }
